squirrel-services/m04-insight-engine/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
sqlalchemy==2.0.21
pymysql==1.1.0
redis==5.0.0
python-dotenv==1.0.0
prometheus-client==0.17.1
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
mypy==1.5.1
